requirements.txt
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
distro==1.9.0
fastapi==0.117.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.11.0
joblib==1.5.2
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
openai==1.109.1
pillow==11.3.0
playwright==1.55.0
playwright-stealth==2.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.4.1
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
